Move label collection class to labels.py
