+++ backend/auto_labeling/pipeline/execution.py
@@ -1,5 +1,4 @@
-import abc
-from typing import List, Type
+from typing import Type
 
 from auto_labeling_pipeline.labels import SequenceLabels, Seq2seqLabels, ClassificationLabels, Labels
 from auto_labeling_pipeline.mappings import MappingTemplate
@@ -7,9 +6,7 @@ from auto_labeling_pipeline.models import RequestModelFactory
 from auto_labeling_pipeline.pipeline import pipeline
 from auto_labeling_pipeline.postprocessing import PostProcessor
 
-from api.models import Example, Project, User
-from api.models import CategoryType, SpanType
-from api.models import Annotation, Category, Span, TextLabel
+from .labels import create_labels
 
 
 def get_label_collection(task_type: str) -> Type[Labels]:
@@ -18,63 +15,6 @@ def get_label_collection(task_type: str) -> Type[Labels]:
         'Span': SequenceLabels,
         'Text': Seq2seqLabels
     }[task_type]
-
-
-class LabelCollection(abc.ABC):
-    label_type = None
-    model = None
-
-    def __init__(self, labels):
-        self.labels = labels
-
-    def transform(self, project: Project, example: Example, user: User) -> List[Annotation]:
-        mapping = {
-            c.text: c for c in self.label_type.objects.filter(project=project)
-        }
-        annotations = []
-        for label in self.labels:
-            if label['label'] not in mapping:
-                continue
-            label['example'] = example
-            label['label'] = mapping[label['label']]
-            label['user'] = user
-            annotations.append(self.model(**label))
-        return annotations
-
-    def save(self, project: Project, example: Example, user: User):
-        labels = self.transform(project, example, user)
-        labels = self.model.objects.filter_annotatable_labels(labels, project)
-        self.model.objects.bulk_create(labels)
-
-
-class Categories(LabelCollection):
-    label_type = CategoryType
-    model = Category
-
-
-class Spans(LabelCollection):
-    label_type = SpanType
-    model = Span
-
-
-class Texts(LabelCollection):
-    model = TextLabel
-
-    def transform(self, project: Project, example: Example, user: User) -> List[Annotation]:
-        annotations = []
-        for label in self.labels:
-            label['example'] = example
-            label['user'] = user
-            annotations.append(self.model(**label))
-        return annotations
-
-
-def create_labels(task_type: str, labels: Labels) -> LabelCollection:
-    return {
-        'Category': Categories,
-        'Span': Spans,
-        'Text': Texts
-    }[task_type](labels.dict())
 
 
 def execute_pipeline(text: str,
