+++ backend/auto_labeling/tests/test_views.py
@@ -10,7 +10,7 @@ from rest_framework.reverse import reverse
 from api.models import DOCUMENT_CLASSIFICATION, SEQUENCE_LABELING, SEQ2SEQ
 from api.models import Category, Span, TextLabel
 from api.tests.api.utils import CRUDMixin, make_doc, prepare_project
-from auto_labeling.pipeline.execution import Categories, Spans, Texts
+from auto_labeling.pipeline.labels import Categories, Spans, Texts
 
 data_dir = pathlib.Path(__file__).parent / 'data'
